fix(bridge-ui): alpha-2 ui fixes (#13395)
