+++ packages/bridge-ui/src/components/TransactionDetail.svelte
@@ -74,15 +74,18 @@
       </tr>
       <tr>
         <td>Data</td>
-        <td
-          class="text-center overflow-auto bg-black !px-1 border-x-4 border-black">
-          {transaction.message.data}
+        <td class="!p-0 border-x-4 tx-data text-center">
+          <div class="h-[50px] leading-[50px] overflow-auto">
+            {transaction.message.data}
+          </div>
         </td>
       </tr>
       <tr>
         <td>Memo</td>
-        <td class="text-right overflow-auto">
-          {transaction.message.memo}
+        <td class="text-right">
+          <div class="overflow-auto">
+            {transaction.message.memo}
+          </div>
         </td>
       </tr>
     {/if}
